+++ src/__tests__/helpers.ts
@@ -45,3 +45,23 @@ export const createFundedUplink = (api: SwitchApi) => async (
 
   return uplink
 }
+
+// Helper that runs deposit/withdraw, capturing and returning the reported tx value and fees.
+export const captureFeesFrom = (apiMethod: any) => async (params: {
+  readonly uplink: ReadyUplinks
+  readonly amount?: BigNumber
+}) => {
+  const reportedValueAndFee = { value: new BigNumber(0), fee: new BigNumber(0) }
+
+  const authorize: AuthorizeDeposit | AuthorizeWithdrawal = async params => {
+    reportedValueAndFee.value = params.value
+    reportedValueAndFee.fee = params.fee
+  }
+
+  await apiMethod({
+    ...params,
+    authorize: authorize
+  })
+
+  return reportedValueAndFee
+}
